+++ packages/filter/src/lib/components/sexpandedFilterGroup/expandedFilterGroup.styles.ts
